+++ src/main/java/janggi/piece/Sang.java
@@ -1,5 +1,6 @@
 package janggi.piece;
 
+import janggi.direction.SangDirection;
 import janggi.setting.CampType;
 import janggi.value.Position;
 import java.util.List;
